+++ test/index_test.js
@@ -163,6 +163,40 @@ describe("Objectware", function() {
     })
   })
 
+  describe(".create", function() {
+    it("must return an object inheriting from the one given", function() {
+      var prototype = {}
+      var obj = _.create(prototype)
+      Object.getPrototypeOf(obj).must.equal(prototype)
+    })
+
+    it("must return an object inheriting from null if given", function() {
+      var obj = _.create(null)
+      demand(Object.getPrototypeOf(obj)).be.null()
+    })
+
+    it("must assign properties to target from one source", function() {
+      _.create({}, {name: "John"}).must.eql({name: "John"})
+    })
+
+    it("must assign properties to target from two sources", function() {
+      var obj = _.create({}, {name: "John"}, {age: 13})
+      obj.must.eql({name: "John", age: 13})
+    })
+
+    it("must throw TypeError given nothing", function() {
+      var err
+      try { _.create() } catch (ex) { err = ex }
+      err.must.be.an.instanceof(TypeError)
+    })
+
+    it("must not modify the given prototype", function() {
+      var prototype = {name: ""}
+      _.create(prototype, {name: "John"})
+      prototype.must.eql({name: ""})
+    })
+  })
+
   describe(".each", function() {
     it("must call function with value, key and object", function() {
       var obj = {name: "John", age: 42, height: 190}